Refactor block architecture into modular package
