+++ project_name/utils/blocks/content.py
@@ -1,0 +1,15 @@
+from wagtail import blocks
+from wagtail.images.blocks import ImageChooserBlock
+
+
+class CaptionedImageBlock(blocks.StructBlock):
+    image = ImageChooserBlock()
+    image_alt_text = blocks.CharBlock(
+        required=False,
+        help_text="If left blank, the image's global alt text will be used.",
+    )
+    caption = blocks.CharBlock(required=False)
+
+    class Meta:
+        icon = "image"
+        template = "components/streamfield/blocks/image_block.html"
